Move RedirectWrapper next to Routes.tsx
