+++ ui/src/Routes.tsx
@@ -20,7 +20,7 @@ import {BrowserRouter as Router, Navigate, Route, Routes as ReactRoutes} from 'r
 import LandingPage from './LandingPage/LandingPage';
 import {OAuthRedirect} from './Common/OAuthRedirect/OAuthRedirect';
 import {AuthenticatedRoute} from './Auth/AuthenticatedRoute';
-import RedirectWrapper from './Common/RedirectWrapper/RedirectWrapper';
+import RedirectWrapper from './RedirectWrapper';
 import SpaceDashboard from './SpaceDashboard/SpaceDashboard';
 import AuthorizedRoute from './Auth/AuthorizedRoute';
 import PeopleMover from './PeopleMover/PeopleMover';
